Add RouteModel
